@@ -1,17 +1,23 @@
 package org.mozilla.tryfox.data
 
+import kotlinx.datetime.Clock
+import kotlinx.datetime.DateTimeUnit
+import kotlinx.datetime.LocalDate
+import kotlinx.datetime.LocalDateTime
+import kotlinx.datetime.TimeZone
+import kotlinx.datetime.format.char
+import kotlinx.datetime.minus
+import kotlinx.datetime.todayIn
 import org.mozilla.tryfox.model.ParsedNightlyApk
 import org.mozilla.tryfox.network.ApiService
 import org.mozilla.tryfox.util.FENIX
 import org.mozilla.tryfox.util.FOCUS
-import java.time.LocalDateTime
-import java.time.format.DateTimeFormatter
+import retrofit2.HttpException
 import java.util.regex.Pattern
-import java.util.Calendar
-import java.util.Locale
 
 class MozillaArchiveRepositoryImpl(
-    private val archiveApiService: ApiService
+    private val archiveApiService: ApiService,
+    private val clock: Clock = Clock.System,
 ) : MozillaArchiveRepository {
 
     companion object {
@@ -19,29 +25,30 @@ class MozillaArchiveRepositoryImpl(
         private const val REFERENCE_BROWSER_TASK_BASE_URL = "https://firefox-ci-tc.services.mozilla.com/api/index/v1/task/mobile.v2.reference-browser.nightly.latest."
         private val REFERENCE_BROWSER_ABIS = listOf("arm64-v8a", "armeabi-v7a", "x86_64")
 
-        internal val fenixArchiveUrl: String get() = archiveUrl(FENIX)
-        internal val focusArchiveUrl: String get() = archiveUrl(FOCUS)
-
-        private fun archiveUrl(appName: String): String {
-            val calendar = Calendar.getInstance()
-            val year = calendar.get(Calendar.YEAR).toString()
-            val month = String.format(
-                Locale.US,
-                "%02d",
-                calendar.get(Calendar.MONTH) + 1
-            ) // Month is 0-indexed
+        internal fun archiveUrlForDate(appName: String, date: LocalDate): String {
+            val year = date.year.toString()
+            val month = date.monthNumber.toString().padStart(2, '0')
 
             return "${ARCHIVE_MOZILLA_BASE_URL}pub/$appName/nightly/$year/$month/"
         }
     }
 
-    override suspend fun getFenixNightlyBuilds(): NetworkResult<List<ParsedNightlyApk>> {
-        return fetchAndParseNightlyBuilds(fenixArchiveUrl, "fenix")
+    private suspend fun getNightlyBuilds(appName: String): NetworkResult<List<ParsedNightlyApk>> {
+        val today = clock.todayIn(TimeZone.currentSystemDefault())
+        val currentMonthUrl = archiveUrlForDate(appName, today)
+        val result = fetchAndParseNightlyBuilds(currentMonthUrl, appName)
+
+        if (result is NetworkResult.Error && (result.cause as? HttpException)?.code() == 404) {
+            val lastMonth = today.minus(1, DateTimeUnit.MONTH)
+            val lastMonthUrl = archiveUrlForDate(appName, lastMonth)
+            return fetchAndParseNightlyBuilds(lastMonthUrl, appName)
+        }
+        return result
     }
 
-    override suspend fun getFocusNightlyBuilds(): NetworkResult<List<ParsedNightlyApk>> {
-        return fetchAndParseNightlyBuilds(focusArchiveUrl, "focus")
-    }
+    override suspend fun getFenixNightlyBuilds(): NetworkResult<List<ParsedNightlyApk>> = getNightlyBuilds(FENIX)
+
+    override suspend fun getFocusNightlyBuilds(): NetworkResult<List<ParsedNightlyApk>> = getNightlyBuilds(FOCUS)
 
     override suspend fun getReferenceBrowserNightlyBuilds(): NetworkResult<List<ParsedNightlyApk>> {
         return try {
@@ -79,28 +86,16 @@ class MozillaArchiveRepositoryImpl(
         archiveUrl: String, // This parameter is the fully constructed URL for the month
         app: String // "fenix" or "focus"
     ): List<ParsedNightlyApk> {
-        val rawBuildStrings = mutableListOf<String>()
-        val htmlPattern =
-            Pattern.compile("""<td>Dir</td>\s*<td><a href="[^"]*">([^<]+/)</a></td>""")
-        val htmlMatcher = htmlPattern.matcher(html)
-        while (htmlMatcher.find()) {
-            htmlMatcher.group(1)?.let {
-                if (it != "../") {
-                    rawBuildStrings.add(it)
-                }
-            }
-        }
+        val htmlPattern = Regex("""<td>Dir</td>\s*<td><a href="[^"]*">([^<]+/)</a></td>""")
+        val rawBuildStrings = htmlPattern.findAll(html)
+            .mapNotNull { it.groups[1]?.value }
+            .filter { it != "../" }
+            .toList()
 
         val buildsByDate = rawBuildStrings.groupBy { it.substringBefore("-$app") }
         if (buildsByDate.isEmpty()) return emptyList()
 
-        val lastBuildDateStr = buildsByDate.keys.maxByOrNull { dateStr ->
-            try {
-                LocalDateTime.parse(dateStr, DateTimeFormatter.ofPattern("yyyy-MM-dd-HH-mm-ss"))
-            } catch (e: Exception) {
-                LocalDateTime.MIN
-            }
-        } ?: return emptyList()
+        val lastBuildDateStr = buildsByDate.keys.maxByOrNull(::parseDateString) ?: return emptyList()
 
         val lastBuildsForDate = buildsByDate[lastBuildDateStr] ?: return emptyList()
 
@@ -128,6 +123,20 @@ class MozillaArchiveRepositoryImpl(
                     fileName = fileName
                 )
             } else null
+        }
+    }
+
+    private fun parseDateString(dateStr: String): LocalDateTime {
+        return try {
+            val format = LocalDateTime.Format {
+                year(); char('-'); monthNumber(); char('-'); dayOfMonth()
+                char('-'); hour(); char('-'); minute(); char('-'); second()
+            }
+            LocalDateTime.parse(dateStr, format)
+
+        } catch (_: Exception) {
+            // Return a very old date so that it's never chosen as the max
+            LocalDateTime(1, 1, 1, 0, 0)
         }
     }
 }
