@@ -2,20 +2,43 @@ package org.mozilla.tryfox.data
 
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
-import org.junit.Assert.assertEquals
-import org.junit.Assert.assertTrue
-import org.junit.Before
-import org.junit.Test
-import org.junit.runner.RunWith
+import kotlinx.datetime.Clock
+import kotlinx.datetime.DateTimeUnit
+import kotlinx.datetime.Instant
+import kotlinx.datetime.LocalDate
+import kotlinx.datetime.TimeZone
+import kotlinx.datetime.atStartOfDayIn
+import kotlinx.datetime.minus
+import okhttp3.ResponseBody.Companion.toResponseBody
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Assertions.assertTrue
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.extension.ExtendWith
 import org.mockito.Mock
-import org.mockito.junit.MockitoJUnitRunner
+import org.mockito.junit.jupiter.MockitoExtension
+import org.mockito.kotlin.eq
 import org.mockito.kotlin.whenever
 import org.mozilla.tryfox.model.ParsedNightlyApk
 import org.mozilla.tryfox.network.ApiService
+import retrofit2.HttpException
+import retrofit2.Response
 
 @ExperimentalCoroutinesApi
-@RunWith(MockitoJUnitRunner::class)
+@ExtendWith(MockitoExtension::class)
 class MozillaArchiveRepositoryImplTest {
+
+    private class FixedClock(private val instant: Instant) : Clock {
+        override fun now(): Instant = instant
+    }
+
+    private companion object {
+        private const val YEAR = "2023"
+        private const val MONTH = "10"
+        private const val FENIX = "fenix"
+        private const val FOCUS = "focus"
+        private const val DATE = "$YEAR-$MONTH-01-10-00-00"
+    }
 
     @Mock
     private lateinit var mockApiService: ApiService
@@ -23,32 +46,27 @@ class MozillaArchiveRepositoryImplTest {
     private lateinit var repository: MozillaArchiveRepositoryImpl
 
     // Test constants for Fenix
-    private val fenixAppName = "fenix"
     private val fenixVersion = "125.0a1"
-    private val fenixDate1 = "2023-11-01-10-00-00"
-    private val fenixAbi1 = "arm64-v8a"
-    private val fenixFileName1 = "$fenixAppName-$fenixVersion.multi.android-$fenixAbi1.apk"
-    private val fenixDirString1 = "$fenixDate1-$fenixAppName-$fenixVersion-android-$fenixAbi1/"
-    private val fenixFullUrl1 = "${MozillaArchiveRepositoryImpl.fenixArchiveUrl}$fenixDirString1$fenixFileName1"
+    private val fenixFileName1 = "$FENIX-$fenixVersion.multi.android-arm64-v8a.apk"
+    private val fenixDirString1 = "$DATE-$FENIX-$fenixVersion-android-arm64-v8a/"
+    private val fenixFullUrl1 = "https://archive.mozilla.org/pub/fenix/nightly/$YEAR/$MONTH/$fenixDirString1$fenixFileName1"
 
-    private val fenixDate2 = "2023-11-01-10-00-00" // Same date, different ABI
-    private val fenixAbi2 = "x86_64"
-    private val fenixFileName2 = "$fenixAppName-$fenixVersion.multi.android-$fenixAbi2.apk"
-    private val fenixDirString2 = "$fenixDate2-$fenixAppName-$fenixVersion-android-$fenixAbi2/"
-    private val fenixFullUrl2 = "${MozillaArchiveRepositoryImpl.fenixArchiveUrl}$fenixDirString2$fenixFileName2"
+    private val fenixFileName2 = "$FENIX-$fenixVersion.multi.android-x86_64.apk"
+    private val fenixDirString2 = "$DATE-$FENIX-$fenixVersion-android-x86_64/"
+    private val fenixFullUrl2 = "https://archive.mozilla.org/pub/fenix/nightly/$YEAR/$MONTH/$fenixDirString2$fenixFileName2"
 
     // Test constants for Focus
-    private val focusAppName = "focus"
     private val focusVersion = "110.0a1"
-    private val focusDate = "2023-10-30-05-30-00"
     private val focusAbi = "armeabi-v7a"
-    private val focusFileName = "$focusAppName-$focusVersion.multi.android-$focusAbi.apk"
-    private val focusDirString = "$focusDate-$focusAppName-$focusVersion-android-$focusAbi/"
-    private val focusFullUrl = "${MozillaArchiveRepositoryImpl.focusArchiveUrl}$focusDirString$focusFileName"
+    private val focusFileName = "$FOCUS-$focusVersion.multi.android-$focusAbi.apk"
+    private val focusDirString = "$DATE-$FOCUS-$focusVersion-android-$focusAbi/"
+    private val focusFullUrl = "https://archive.mozilla.org/pub/focus/nightly/$YEAR/$MONTH/$focusDirString$focusFileName"
 
-    @Before
+    @BeforeEach
     fun setUp() {
-        repository = MozillaArchiveRepositoryImpl(mockApiService)
+        val testDate = LocalDate(2023, 10, 1)
+        val clock = FixedClock(testDate.atStartOfDayIn(TimeZone.UTC))
+        repository = MozillaArchiveRepositoryImpl(mockApiService, clock)
     }
 
     private fun createMockHtmlResponse(vararg dirStrings: String): String {
@@ -60,7 +78,8 @@ class MozillaArchiveRepositoryImplTest {
     @Test
     fun `getFenixNightlyBuilds success - single latest build`() = runTest {
         val mockHtml = createMockHtmlResponse(fenixDirString1)
-        whenever(mockApiService.getHtmlPage(MozillaArchiveRepositoryImpl.fenixArchiveUrl)).thenReturn(mockHtml)
+        val expectedUrl = "https://archive.mozilla.org/pub/fenix/nightly/2023/10/"
+        whenever(mockApiService.getHtmlPage(eq(expectedUrl))).thenReturn(mockHtml)
 
         val result = repository.getFenixNightlyBuilds()
 
@@ -69,19 +88,20 @@ class MozillaArchiveRepositoryImplTest {
         assertEquals(1, apks.size)
         val apk = apks[0]
         assertEquals(fenixDirString1, apk.originalString)
-        assertEquals(fenixDate1, apk.rawDateString)
-        assertEquals(fenixAppName, apk.appName)
+        assertEquals(DATE, apk.rawDateString)
+        assertEquals(FENIX, apk.appName)
         assertEquals(fenixVersion, apk.version)
-        assertEquals(fenixAbi1, apk.abiName)
+        assertEquals("arm64-v8a", apk.abiName)
         assertEquals(fenixFullUrl1, apk.fullUrl)
         assertEquals(fenixFileName1, apk.fileName)
     }
 
     @Test
     fun `getFenixNightlyBuilds success - multiple builds on latest date`() = runTest {
-        val olderDateDir = "2023-10-31-00-00-00-$fenixAppName-$fenixVersion-android-$fenixAbi1/"
+        val olderDateDir = "2023-9-31-00-00-00-$FENIX-$fenixVersion-android-arm64-v8a/"
         val mockHtml = createMockHtmlResponse(fenixDirString1, fenixDirString2, olderDateDir)
-        whenever(mockApiService.getHtmlPage(MozillaArchiveRepositoryImpl.fenixArchiveUrl)).thenReturn(mockHtml)
+        val expectedUrl = "https://archive.mozilla.org/pub/fenix/nightly/2023/10/"
+        whenever(mockApiService.getHtmlPage(eq(expectedUrl))).thenReturn(mockHtml)
 
         val result = repository.getFenixNightlyBuilds()
 
@@ -90,20 +110,22 @@ class MozillaArchiveRepositoryImplTest {
         assertEquals(2, apks.size)
 
         val expectedApks = listOf(
-            ParsedNightlyApk(fenixDirString1, fenixDate1, fenixAppName, fenixVersion, fenixAbi1, fenixFullUrl1, fenixFileName1),
-            ParsedNightlyApk(fenixDirString2, fenixDate2, fenixAppName, fenixVersion, fenixAbi2, fenixFullUrl2, fenixFileName2)
+            ParsedNightlyApk(fenixDirString1, DATE, FENIX, fenixVersion, "arm64-v8a", fenixFullUrl1, fenixFileName1),
+            ParsedNightlyApk(fenixDirString2, DATE, FENIX, fenixVersion, "x86_64", fenixFullUrl2, fenixFileName2)
         )
+        assertEquals(expectedApks.first(), apks.first())
         assertTrue(apks.containsAll(expectedApks) && expectedApks.containsAll(apks))
     }
 
      @Test
     fun `getFenixNightlyBuilds success - sorts by date correctly`() = runTest {
-        val olderDateDir = "2023-10-31-23-59-59-$fenixAppName-$fenixVersion-android-$fenixAbi1/"
-        val latestDateDir = fenixDirString1
-        val middleDateDir = "2023-11-01-09-00-00-$fenixAppName-$fenixVersion-android-x86/"
+        val olderDateDir = "2023-09-31-23-59-59-$FENIX-$fenixVersion-android-arm64-v8a/"
+        val middleDateDir = fenixDirString1
+        val latestDateDir = "2023-11-01-09-00-00-$FENIX-$fenixVersion-android-x86/"
 
         val mockHtml = createMockHtmlResponse(olderDateDir, latestDateDir, middleDateDir)
-        whenever(mockApiService.getHtmlPage(MozillaArchiveRepositoryImpl.fenixArchiveUrl)).thenReturn(mockHtml)
+        val expectedUrl = "https://archive.mozilla.org/pub/fenix/nightly/2023/10/"
+        whenever(mockApiService.getHtmlPage(eq(expectedUrl))).thenReturn(mockHtml)
 
         val result = repository.getFenixNightlyBuilds()
         assertTrue(result is NetworkResult.Success)
@@ -115,7 +137,8 @@ class MozillaArchiveRepositoryImplTest {
     @Test
     fun `getFenixNightlyBuilds success - no builds found`() = runTest {
         val mockHtml = "<td>Some other HTML</td>"
-        whenever(mockApiService.getHtmlPage(MozillaArchiveRepositoryImpl.fenixArchiveUrl)).thenReturn(mockHtml)
+        val expectedUrl = "https://archive.mozilla.org/pub/fenix/nightly/2023/10/"
+        whenever(mockApiService.getHtmlPage(eq(expectedUrl))).thenReturn(mockHtml)
 
         val result = repository.getFenixNightlyBuilds()
 
@@ -126,7 +149,8 @@ class MozillaArchiveRepositoryImplTest {
     @Test
     fun `getFenixNightlyBuilds network error`() = runTest {
         val errorMessage = "Network error"
-        whenever(mockApiService.getHtmlPage(MozillaArchiveRepositoryImpl.fenixArchiveUrl)).thenThrow(RuntimeException(errorMessage))
+        val expectedUrl = "https://archive.mozilla.org/pub/fenix/nightly/2023/10/"
+        whenever(mockApiService.getHtmlPage(eq(expectedUrl))).thenThrow(RuntimeException(errorMessage))
 
         val result = repository.getFenixNightlyBuilds()
 
@@ -135,9 +159,41 @@ class MozillaArchiveRepositoryImplTest {
     }
 
     @Test
+    fun `getFenixNightlyBuilds when current month 404 queries previous month`() = runTest {
+        // Given
+        val testDate = LocalDate(2024, 3, 15)
+        val clock = FixedClock(testDate.atStartOfDayIn(TimeZone.UTC))
+        val repository = MozillaArchiveRepositoryImpl(mockApiService, clock)
+
+        val currentMonthUrl = MozillaArchiveRepositoryImpl.archiveUrlForDate(FENIX, testDate)
+
+        val previousMonthDate = testDate.minus(1, DateTimeUnit.MONTH)
+        val previousMonthUrl = MozillaArchiveRepositoryImpl.archiveUrlForDate(FENIX, previousMonthDate)
+
+        val previousMonthDateString = "2024-02-28-10-00-00"
+        val previousMonthDirString = "$previousMonthDateString-$FENIX-$fenixVersion-android-arm64-v8a/"
+        val previousMonthMockHtml = createMockHtmlResponse(previousMonthDirString)
+
+        whenever(mockApiService.getHtmlPage(currentMonthUrl)).thenThrow(
+            HttpException(Response.error<Any>(404, "".toResponseBody(null)))
+        )
+        whenever(mockApiService.getHtmlPage(previousMonthUrl)).thenReturn(previousMonthMockHtml)
+
+        // When
+        val result = repository.getFenixNightlyBuilds()
+
+        // Then
+        assertTrue(result is NetworkResult.Success)
+        val apks = (result as NetworkResult.Success).data
+        assertEquals(1, apks.size)
+        assertEquals(previousMonthDirString, apks[0].originalString)
+    }
+
+    @Test
     fun `getFocusNightlyBuilds success - single latest build`() = runTest {
         val mockHtml = createMockHtmlResponse(focusDirString)
-        whenever(mockApiService.getHtmlPage(MozillaArchiveRepositoryImpl.focusArchiveUrl)).thenReturn(mockHtml)
+        val expectedUrl = "https://archive.mozilla.org/pub/focus/nightly/2023/10/"
+        whenever(mockApiService.getHtmlPage(eq(expectedUrl))).thenReturn(mockHtml)
 
         val result = repository.getFocusNightlyBuilds()
 
@@ -146,8 +202,8 @@ class MozillaArchiveRepositoryImplTest {
         assertEquals(1, apks.size)
         val apk = apks[0]
         assertEquals(focusDirString, apk.originalString)
-        assertEquals(focusDate, apk.rawDateString)
-        assertEquals(focusAppName, apk.appName)
+        assertEquals(DATE, apk.rawDateString)
+        assertEquals(FOCUS, apk.appName)
         assertEquals(focusVersion, apk.version)
         assertEquals(focusAbi, apk.abiName)
         assertEquals(focusFullUrl, apk.fullUrl)
@@ -157,7 +213,8 @@ class MozillaArchiveRepositoryImplTest {
     @Test
     fun `getFocusNightlyBuilds network error`() = runTest {
         val errorMessage = "Network error for Focus"
-        whenever(mockApiService.getHtmlPage(MozillaArchiveRepositoryImpl.focusArchiveUrl)).thenThrow(RuntimeException(errorMessage))
+        val expectedUrl = "https://archive.mozilla.org/pub/focus/nightly/2023/10/"
+        whenever(mockApiService.getHtmlPage(eq(expectedUrl))).thenThrow(RuntimeException(errorMessage))
 
         val result = repository.getFocusNightlyBuilds()
 
