@@ -6,21 +6,20 @@ import kotlinx.coroutines.test.TestDispatcher
 import kotlinx.coroutines.test.StandardTestDispatcher
 import kotlinx.coroutines.test.resetMain
 import kotlinx.coroutines.test.setMain
-import org.junit.rules.TestWatcher
-import org.junit.runner.Description
+import org.junit.jupiter.api.extension.AfterEachCallback
+import org.junit.jupiter.api.extension.BeforeEachCallback
+import org.junit.jupiter.api.extension.ExtensionContext
 
 @ExperimentalCoroutinesApi
 class MainCoroutineRule(
-    val testDispatcher: TestDispatcher = StandardTestDispatcher()
-) : TestWatcher() {
+    val testDispatcher: TestDispatcher = StandardTestDispatcher(),
+) : BeforeEachCallback, AfterEachCallback {
 
-    override fun starting(description: Description) {
-        super.starting(description)
+    override fun beforeEach(context: ExtensionContext) {
         Dispatchers.setMain(testDispatcher)
     }
 
-    override fun finished(description: Description) {
-        super.finished(description)
+    override fun afterEach(context: ExtensionContext) {
         Dispatchers.resetMain()
     }
 }
